Better json transcription
the words are now formed on the "title_orthography" part of json. this is because the key sometimes is modified due to duplicates, leading to the words formed being weird.

this however leads the unicode escape character problem to be at the forefront but has been solved thanks to :https://stackoverflow.com/questions/18337407/saving-utf-8-texts-with-json-dumps-as-utf-8-not-as-a-u-escape-sequence (the ensure_ascii=False)

now all accents are present in the output json